# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktaba Dot Com - Arabic RTL E-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Admin product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductListResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
  /admin/products/{id}:
    get:
      operationId: adminGetProduct
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Admin product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/products/{id}/stock:
    patch:
      operationId: adminUpdateProductStock
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockUpdateInput"
      responses:
        "200":
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"

  # ─── ADMIN – ORDERS ──────────────────────────────────────────────────────────
...
  /admin/stages:
    get:
      operationId: adminListStages
      tags: [admin]
      responses:
        "200":
          description: Stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stage"
    post:
      operationId: adminCreateStage
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageInput"
      responses:
        "201":
          description: Stage created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
  /admin/stages/{id}:
    patch:
      operationId: adminUpdateStage
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageUpdate"
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
    delete:
      operationId: adminDeleteStage
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/grades:
    get:
      operationId: adminListGrades
      tags: [admin]
      responses:
        "200":
          description: Grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: adminCreateGrade
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: Grade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /admin/grades/{id}:
    patch:
      operationId: adminUpdateGrade
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeUpdate"
      responses:
        "200":
          description: Updated grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: adminDeleteGrade
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/subjects:
    get:
      operationId: adminListSubjects
      tags: [admin]
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: adminCreateSubject
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /admin/subjects/{id}:
    patch:
      operationId: adminUpdateSubject
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: adminDeleteSubject
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /admin/publishers:
    get:
      operationId: adminListPublishers
      tags: [admin]
      responses:
        "200":
          description: Publishers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publisher"
    post:
      operationId: adminCreatePublisher
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherInput"
      responses:
        "201":
          description: Publisher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
  /admin/publishers/{id}:
    patch:
      operationId: adminUpdatePublisher
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherUpdate"
      responses:
        "200":
          description: Updated publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
    delete:
      operationId: adminDeletePublisher
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ADMIN – COUPONS ─────────────────────────────────────────────────────────
...
  /admin/stock-movements:
    get:
      operationId: adminListStockMovements
      tags: [admin]
      parameters:
        - in: query
          name: productId
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovementListResponse"

# ─── COMPONENTS ──────────────────────────────────────────────────────────────
components:
  schemas:
...
    Stage:
      type: object
      required: [id, nameAr]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    StageInput:
      type: object
      required: [nameAr]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        sortOrder: { type: integer }
    StageUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    Grade:
      type: object
      required: [id, nameAr, stageId]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        stageId: { type: integer }
        stageName: { type: ["string", "null"] }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    GradeInput:
      type: object
      required: [nameAr, stageId]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        stageId: { type: integer }
        sortOrder: { type: integer }
    GradeUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        stageId: { type: integer }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    Subject:
      type: object
      required: [id, nameAr]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        isActive: { type: boolean }
    SubjectInput:
      type: object
      required: [nameAr]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
    SubjectUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        isActive: { type: boolean }
    Publisher:
      type: object
      required: [id, nameAr]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        logo: { type: ["string", "null"] }
        isActive: { type: boolean }
    PublisherInput:
      type: object
      required: [nameAr]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        logo: { type: string }
    PublisherUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        logo: { type: string }
        isActive: { type: boolean }

    # ── Shipping / Governorates ───────────────────────────────────────────────
...
    AdminProduct:
      type: object
      required: [id, nameAr, slug, price, status, stockQuantity]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: ["string", "null"] }
        slug: { type: string }
        descriptionShort: { type: ["string", "null"] }
        descriptionFull: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        price: { type: number }
        oldPrice: { type: ["number", "null"] }
        sku: { type: ["string", "null"] }
        barcode: { type: ["string", "null"] }
        status: { type: string, enum: [active, draft, archived] }
        stockQuantity: { type: integer }
        reservedQuantity: { type: integer }
        minStockLevel: { type: integer }
        stageId: { type: ["integer", "null"] }
        gradeId: { type: ["integer", "null"] }
        subjectId: { type: ["integer", "null"] }
        publisherId: { type: ["integer", "null"] }
        educationType: { type: ["string", "null"] }
        bookType: { type: ["string", "null"] }
        edition: { type: ["string", "null"] }
        schoolYear: { type: ["string", "null"] }
        isBestSeller: { type: boolean }
        isFeatured: { type: boolean }
        isNew: { type: boolean }
        isRevision: { type: boolean }
        isBundle: { type: boolean }
        sortOrder: { type: integer }
        internalNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AdminProductInput:
      type: object
      required: [nameAr, price]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionShort: { type: string }
        descriptionFull: { type: string }
        coverImage: { type: string }
        images: { type: array, items: { type: string } }
        price: { type: number }
        oldPrice: { type: number }
        sku: { type: string }
        barcode: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        stockQuantity: { type: integer }
        minStockLevel: { type: integer }
        stageId: { type: integer }
        gradeId: { type: integer }
        subjectId: { type: integer }
        publisherId: { type: integer }
        educationType: { type: string }
        bookType: { type: string }
        edition: { type: string }
        schoolYear: { type: string }
        isBestSeller: { type: boolean }
        isFeatured: { type: boolean }
        isNew: { type: boolean }
        isRevision: { type: boolean }
        isBundle: { type: boolean }
        internalNotes: { type: string }
    AdminProductUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionShort: { type: string }
        descriptionFull: { type: string }
        coverImage: { type: string }
        images: { type: array, items: { type: string } }
        price: { type: number }
        oldPrice: { type: number }
        sku: { type: string }
        barcode: { type: string }
        status: { type: string }
        minStockLevel: { type: integer }
        stageId: { type: integer }
        gradeId: { type: integer }
        subjectId: { type: integer }
        publisherId: { type: integer }
        educationType: { type: string }
        bookType: { type: string }
        edition: { type: string }
        schoolYear: { type: string }
        isBestSeller: { type: boolean }
        isFeatured: { type: boolean }
        isNew: { type: boolean }
        isRevision: { type: boolean }
        isBundle: { type: boolean }
        internalNotes: { type: string }
    AdminProductListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminProduct"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    StockUpdateInput:
      type: object
      required: [quantity, movementType, reason]
      properties:
        quantity: { type: integer }
        movementType: { type: string, enum: [purchase, sale, return, damaged, manual_increase, manual_decrease, adjustment] }
        reason: { type: string }

    # ── Admin Orders ─────────────────────────────────────────────────────────
...
    StockMovement:
      type: object
      required: [id, productId, movementType, quantityBefore, quantityAfter, quantityChanged, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productNameAr: { type: ["string", "null"] }
        movementType: { type: string }
        quantityBefore: { type: integer }
        quantityAfter: { type: integer }
        quantityChanged: { type: integer }
        reason: { type: ["string", "null"] }
        orderId: { type: ["integer", "null"] }
        employeeName: { type: ["string", "null"] }
        createdAt: { type: string }
    StockMovementListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StockMovement"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }