# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Maktaba Dot Com - Arabic RTL E-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface StockMovement {
  id: number;
  productId: number;
  /** @nullable */
  productNameAr?: string | null;
  movementType: string;
  quantityBefore: number;
  quantityAfter: number;
  quantityChanged: number;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  orderId?: number | null;
  /** @nullable */
  employeeName?: string | null;
  createdAt: string;
}